requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.31.0
pdfplumber==0.11.0
Pillow==10.3.0
supabase>=2.0.0
